+++ example/integration_test/session_tests/automatic_sessions/206_CR_CNG_A_id_change_test.dart
@@ -2,7 +2,6 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:countly_flutter_np/countly_flutter.dart';
-import 'package:flutter_foreground_task/flutter_foreground_task.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 import '../../utils.dart';
@@ -20,9 +19,7 @@ void main() {
     await Countly.instance.sessions.updateSession();
     await Countly.instance.sessions.endSession();
 
-    FlutterForegroundTask.minimizeApp();
-    await tester.pump(Duration(seconds: 1));
-    FlutterForegroundTask.launchApp();
+    goBackgroundAndForeground();
 
     await tester.pump(Duration(seconds: 1));
     await Countly.changeDeviceId('newID_2', false);